backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
python-decouple==3.8
requests==2.31.0
qrcode==7.4.2
Pillow==10.1.0
gunicorn==21.2.0
